+++ src/snowflake/snowpark/_internal/analyzer/snowflake_plan.py
@@ -1813,7 +1813,7 @@ class SnowflakePlanBuilder:
         xml_reader_udtf: "UserDefinedTableFunction",
         file_path: str,
         options: Dict[str, str],
-        schema_string: str,
+        schema: List[Attribute],
     ) -> str:
         """
         Creates a DataFrame from a UserDefinedTableFunction that reads XML files.
@@ -1821,6 +1821,7 @@ class SnowflakePlanBuilder:
         from snowflake.snowpark.functions import lit, col, seq8, flatten
         from snowflake.snowpark._internal.xml_reader import DEFAULT_CHUNK_SIZE
 
+        schema_string = attribute_to_schema_string(schema) if schema is not None else ""
         worker_column_name = "WORKER"
         xml_row_number_column_name = "XML_ROW_NUMBER"
         row_tag = options[XML_ROW_TAG_STRING]
@@ -1908,10 +1909,9 @@ class SnowflakePlanBuilder:
         source_plan: Optional[ReadFileNode] = None,
     ) -> SnowflakePlan:
         thread_safe_session_enabled = self.session._conn._thread_safe_session_enabled
-        schema_string = attribute_to_schema_string(schema)
         if xml_reader_udtf is not None:
             xml_query = self._create_xml_query(
-                xml_reader_udtf, path, options, schema_string if use_user_schema else ""
+                xml_reader_udtf, path, options, schema if use_user_schema else None
             )
             return SnowflakePlan(
                 [Query(xml_query)],